Refactor: 피어 리뷰 반영 - (4) Vuex Store 모듈화 (CartModule.js)
